git-svn-id: svn+ssh://jabouzi.com/home/jabouzic/svn/projet/trunk@394 b4f2fc90-cf10-4682-8399-fbf5ccf6d971

+++ app/dao/userdatadao.php
@@ -31,18 +31,21 @@ class Userdatadao {
 	public function insert_vhosts($user)
 	{
 		$insert = 0;
-		foreach($user->get_user_vhosts() as $vhost)
+		if (count($user->get_user_vhosts()))
 		{
-			$args_vhost = array(
-				':user_name' => $user->get_user_name,
-				':vhost' => $host
-			);
-			$query = "INSERT INTO user_vhost VALUES (:user_name, :vhost)";
-			$insert += $this->db->query($query, $args_vhost);
+			foreach($user->get_user_vhosts() as $vhost)
+			{
+				$args_vhost = array(
+					':user_name' => $user->get_user_name,
+					':vhost' => $host
+				);
+				$query = "INSERT INTO user_vhost VALUES (:user_name, :vhost)";
+				$insert += $this->db->query($query, $args_vhost);
+			}
+			$this->cache->delete('select_data_'.$user->get_user_name);
+			$this->cache->delete('select_data_all');
+			return ($insert == count($user->get_vhosts));
 		}
-		$this->cache->delete('select_data_'.$user->get_user_name);
-		$this->cache->delete('select_data_all');
-		return ($insert == count($user->get_vhosts));
 	}
 
 	public function update_info($user)
